check if dist folder exists, then remove content

+++ create-package.sh
@@ -5,7 +5,9 @@ set -e
 source yapf-module.sh
 source flake8-module.sh
 
-rm dist/*
+if [ -d "dist" ]; then
+    rm dist/*
+fi
 pip install --upgrade build
 python3 -m build
 #python3 setup.py bdist_wheel
